New changes to add some logging, simplified implementation. Should not provide any actual new features or bugs.
I hope.

*crosses fingers*
